refactor: split synthetic run writer
